minor refactoring, put classes in files, make compare work better
